:recycle: Refactor data-loading, add celery beat fixture to it
