+++ src/sdg/accounts/apps.py
@@ -1,27 +1,5 @@
-from django.apps import AppConfig, apps
-from django.conf import settings
-from django.contrib.contenttypes.management import create_contenttypes
-from django.core.management import call_command
-from django.db.models.signals import post_migrate
-
-
-def update_admin_index(sender, **kwargs):
-    from django_admin_index.models import AppGroup
-
-    AppGroup.objects.all().delete()
-
-    # Make sure project models are registered.
-    project_name = __name__.split(".")[0]
-
-    for app_config in apps.get_app_configs():
-        if app_config.name.startswith(project_name):
-            create_contenttypes(app_config)
-
-    call_command("loaddata", "default_admin_index", verbosity=0)
+from django.apps import AppConfig
 
 
 class AccountsConfig(AppConfig):
     name = "sdg.accounts"
-
-    def ready(self):
-        post_migrate.connect(update_admin_index, sender=self)
